requirements.txt
fastapi
uvicorn
easyocr
pydantic
pillow
numpy
opencv-python
python-multipart
opencv-python-headless